+++ db/migrate/20160226140743_add_dsmb_member_to_users.rb
@@ -1,0 +1,6 @@
+class AddDsmbMemberToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :dsmb_member, :boolean, null: false, default: false
+    add_index :users, :dsmb_member
+  end
+end